+++ docs/conf.py
@@ -33,9 +33,9 @@ author = "Wildfish"
 # built documents.
 #
 # The short X.Y version.
-version = "0.1"
+version = "0.2"
 # The full version, including alpha/beta/rc tags.
-release = "0.1.0"
+release = "0.2.0"
 
 # -- General configuration ---------------------------------------------------